new multiuser scheme, one lab for three environments
git-svn-id: https://tor.ern.nps.edu/svn/proj/labtainer/trunk@1594 ed162e55-3153-4343-a18d-e086fa355ed4

+++ scripts/labtainer-student/lab_bin/parameterize.sh
@@ -104,8 +104,9 @@ fi
 if [ -f /usr/bin/apt-source.sh ]; then
     echo $CONTAINER_PASSWORD | sudo -S /usr/bin/apt-source.sh
 fi
-
-rmdir $LOCKDIR
+if [ -d $LOCKDIR ]; then
+    rmdir $LOCKDIR
+fi
 # Added a permanent 'did_param' lock directory
 PERMLOCKDIR=/var/labtainer/did_param
 echo $CONTAINER_PASSWORD | sudo -S mkdir -p "$PERMLOCKDIR" >/dev/null 2>&1
